Split imarker_robot_state into library
