+++ hilgendorf_moveit_demos/include/hilgendorf_moveit_demos/hilgendorf_demos.h
@@ -40,7 +40,7 @@
    Visualizations
    - visual_ompl1_ - used by BoltDB as VizCallbacks
    - visual_ompl2_ - used by BoltDB as Viz2Callbacks
-   - visual_ompl3_ - shows the solution robot path and trajectory line
+   - visual_ompl3_ - used by BoltDBRetriveRepair as VizCallbacks, also shows the solution robot path and trajectory line
    - visual_moveit_start_ - same as visual_ompl1_ - shows start state
    - visual_moveit_goal_  - same as visual_ompl2_ - shows goal state
    - visual_tools_ - from moveit_base base class
@@ -152,6 +152,8 @@ public:
   bool visualize_grid_generation_;
   bool visualize_start_goal_states_;
   bool visualize_astar_;
+  bool visualize_cart_neighbors_;
+  bool visualize_cart_path_;
   double visualize_time_between_plans_;
   bool debug_print_trajectory_;
